refactor(engine): extract apply_batch_parallel into parallel_apply/batch.rs (SPL-38.d)
Continues the SPL-38 decomposition of `concurrent_delta/parallel_apply/mod.rs`.
Moves the `apply_batch_parallel` method (and the three tests that exclusively
cover it) into a new sibling `batch.rs`, mirroring the slot_barrier.rs /
decrement_guard.rs precedent. Shared test helpers in `mod.rs::tests` are
promoted to `pub(super)` so the new `batch::tests` module can reuse them.
Public API and behaviour are unchanged.
